Work around the <atom:title> issue
Instead of trying to make the <![CDATA element "safe", pull the entire
start of the <title> tag into the unescaped portion alonw with the start
of the <![CDATA element. This lets html/template think we're just
dropping some mad HTML in there and not some syntax-errorful nonsense.

Feed validator is happy, and I'm mostly content.